added filters

+++ akash_search_app/serializers.py
@@ -7,16 +7,13 @@ class ProductSerializer(serializers.ModelSerializer):
     class Meta:
         model = Product
         fields = '__all__'
-
-
     def to_representation(self, instance):
         representation = super().to_representation(instance)
         for field in ['MRP', 'SellPrice']:
             try:
-                # Convert to Decimal
-                value = Decimal(str(getattr(instance, field, '0')).strip())
+                # Convert to Float
+                value = float((str(getattr(instance, field, '0')).strip()))
                 representation[field] = value
             except (ValueError, InvalidOperation):
-                # Handle the invalid value by setting it to None or a default value
                 representation[field] = None
         return representation
